#1 09 권한 체크

+++ src/main/java/com/ll/ch03_10/domain/surl/surl/controller/ApiV1SurlController.java
@@ -71,6 +71,10 @@ public class ApiV1SurlController {
             @PathVariable long id
     ){
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
+        Member member = rq.getMember();
+        if(!surl.getAuthor().equals(member)){
+            throw new GlobalException("403-1", "권한이 없습니다.");
+        }
 
         return RsData.of(
                 new SurlGetRespBody(
@@ -86,6 +90,12 @@ public class ApiV1SurlController {
             @PathVariable long id
     ){
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
+
+        Member member = rq.getMember();
+        if(!surl.getAuthor().equals(member)){
+            throw new GlobalException("403-1", "권한이 없습니다.");
+        }
+
         surlService.delete(surl);
 
         return RsData.OK;
@@ -139,6 +149,11 @@ public class ApiV1SurlController {
     ){
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
         RsData<Surl> modifyRs = surlService.modify(surl, reqBody.body, reqBody.url);
+
+        Member member = rq.getMember();
+        if(!surl.getAuthor().equals(member)){
+            throw new GlobalException("403-1", "권한이 없습니다.");
+        }
 
         return modifyRs.newDataOf(
                 new SurlModifyRespBody(
